Make the transition to Kotlin, JavaFX and GitLab
This commit is rather a total conversion of the tool into Kotlin
code. There have been many simplifications and many removals of
Swing classes (in order to move to JavaFX).

Apart from switching languages Gradle is used as primary build
system. That means that a `gradle run` compiles and executes the
application while downloading all dependencies.

One of these dependencies is iTextPDF which has been updated to
use version 7 instead of 5 being much faster than the previous
code. An issue which is unsolved at this point is image handling,
because the images are currently scaled according to page size
which is mostly unsuitable.

Another dependency is `kotlin-logging` which serves as logging
framework. Any bugtracker-related and Apache Commons code has
been removed. The ControlsFX framework supports JavaFX for some
nice controls.